Implement non-blocking streaming of updates
Previously, we had to wait for a response before sending
out a batch of updates.

This is not as inefficient as it sounds, since we're batching
multiple updates with each 'raft append entries', so the latency
impact was amortized over many redis commands.

Still, good idea to have true non-blocking streaming. No
change in performance, though, as far I can tell.