Stall readers when necessary, instead of giving an error
There's a small window between a leader getting elected
and its state machine catching up to the leadership marker,
during which it's unsafe to service reads, as there's the
possibility of giving out stale values.

Instead of giving an error to the client, stall the connection
until it's safe to reply.