Implement SCAN command with kinda-efficient pattern matching
The efficiency of the pattern matching depends on how many keys
start with the non-glob prefix of the pattern.

Examples

"abc*" will be very efficient, hitting just the keys
that match.

"*abc" will be inefficient, since we have to search through
the entire keyspace.

"ab?d" will have to iterate through all keys starting with
"ab".