Shard gtest runs on fedora-tsan-test to prevent running out of memory
It appears that ThreadSanitizer does not release memory to the OS,
or re-use the memory after a delete. This makes running the stress
tests difficult, as we run out of memory.

Let's try using a feature from gtest meant for distributed test
execution, which is to shard the tests over different runs. This
way, we will reclaim memory after each "shard" finishes running,
and use far less memory in total.